Move Committee model
