+++ webservices/committees/models.py
@@ -1,0 +1,155 @@
+class Committee(object):
+
+    default_fields = {
+        'dimcmte_fields': 'cmte_id,form_tp,load_date,expire_date,',
+        'properties_fields': '*,',
+        'linkages_fields': (
+            'cand_id,cmte_tp,cmte_dsgn,cand_election_yr,expire_date,'
+            'link_date, /dimcand{/dimcandproperties{cand_nm,}, '
+            '/dimcandoffice{/dimoffice{office_tp,office_tp_desc}}}'),
+        'designation_fields': '*,',
+    }
+
+    table_name_stem = 'cmte'
+    viewable_table_name = "(dimcmte?exists(dimcmteproperties))"
+
+    def query_text(self, show_fields):
+        # We always need expire date and cand_id to sort the information
+        return (
+            '(%s{{%s max(dimcmteproperties.cmte_nm)},'
+            '/dimcmteproperties{expire_date,%s}, '
+            '/dimlinkages{cand_id,expire_date,%s}, '
+            '/dimcmtetpdsgn{expire_date,%s}})'
+            '.sort(max(dimcmteproperties.cmte_nm)+)') % (
+            self.viewable_table_name,
+            show_fields['dimcmte_fields'],
+            show_fields['properties_fields'],
+            show_fields['linkages_fields'],
+            show_fields['designation_fields']
+        )
+
+    dimcmte_mapping = (
+        ('committee_id', 'cmte_id'),
+        ('form_type', 'form_tp'),
+        ('load_date', 'load_date'),
+        ('expire_date', 'expire_date'),
+        ('*', '*'),
+    )
+    linkages_field_mapping = (
+        ('candidate_id', 'cand_id'),
+        ('type', 'cmte_tp'),
+        ('designation', 'cmte_dsgn'),
+        ('election_years', 'cand_election_yr'),
+        ('expire_date', 'expire_date'),
+        ('link_date', 'link_date'),
+    )
+    linkages_mapping = (
+        ('committees', 'cand_id,cmte_tp,cmte_dsgn,cand_election_yr,\
+            expire_date,link_date'),
+        ('candidate_name', '/dimcand{/dimcandproperties{cand_nm}}'),
+        ('office_sought', '/dimcandoffice{/dimoffice{office_tp}}}'),
+        ('office_sought_full', '/dimcandoffice{/dimoffice{office_tp_desc}}}'),
+        ('candidate_name', '/dimcand{/dimcandproperties{cand_nm,}'),
+        ('*', '*, /dimcand{/dimcandproperties{cand_nm,}, '
+              '/dimcandoffice{/dimoffice{office_tp,office_tp_desc}}}'),
+    ) + linkages_field_mapping
+
+    designation_mapping = (
+        ('designation', 'cmte_dsgn'),
+        ('type', 'cmte_tp'),
+        ('expire_date', 'expire_date'),
+        ('load_date', 'load_date'),
+        ('receipt_date', 'receipt_date'),
+        ('*', '*'),
+    )
+
+    # Beginning of the dimcmteproperties mapping, making subdivisions for
+    # easier formatting
+    committee_address_field_mappings = (
+        ('street_1', 'cmte_st1'),
+        ('street_2', 'cmte_st2'),
+        ('city', 'cmte_city'),
+        ('state', 'cmte_st'),
+        ('state_full', 'cmte_st_desc'),
+        ('zip', 'cmte_zip'),
+    )
+    committee_address = (
+        'address',
+        'street_1,cmte_st1,street_2,cmte_st2,cmte_city,cmte_st,cmte_zip')
+
+    treasurer_field_mapping = (
+        ('city', 'cmte_treasurer_city'),
+        ('name_1', 'cmte_treasurer_f_nm'),
+        ('name_2', 'cmte_treasurer_l_nm'),
+        ('name_middle', 'cmte_treasurer_m_nm'),
+        ('name_full', 'cmte_treasurer_nm'),
+        ('phone', 'cmte_treasurer_ph_num'),
+        ('name_prefix', 'cmte_treasurer_prefix'),
+        ('state', 'cmte_treasurer_st'),
+        ('street_1', 'cmte_treasurer_st1'),
+        ('street_2', 'cmte_treasurer_st2'),
+        ('name_suffix', 'cmte_treasurer_suffix'),
+        ('name_title', 'cmte_treasurer_title'),
+        ('zip', 'cmte_treasurer_zip'),
+    )
+    treasurer = (
+        ('treasurer', 'cmte_treasurer_city,cmte_treasurer_f_nm,\
+            cmte_treasurer_l_nm,cmte_treasurer_m_nm,cmte_treasurer_nm,\
+            cmte_treasurer_ph_num,cmte_treasurer_prefix,cmte_treasurer_st,\
+            cmte_treasurer_st1,cmte_treasurer_st2,cmte_treasurer_suffix,\
+            cmte_treasurer_title,cmte_treasurer_zip'),
+    )
+    custodian_field_mapping = (
+        ('city', 'cmte_custodian_city'),
+        ('name_1', 'cmte_custodian_f_nm'),
+        ('name_2', 'cmte_custodian_l_nm'),
+        ('name_middle', 'cmte_custodian_m_nm'),
+        ('name_full', 'cmte_custodian_nm'),
+        ('phone', 'cmte_custodian_ph_num'),
+        ('name_prefix', 'cmte_custodian_prefix'),
+        ('state', 'cmte_custodian_st'),
+        ('street_1', 'cmte_custodian_st1'),
+        ('street_2', 'cmte_custodian_st2'),
+        ('name_suffix', 'cmte_custodian_suffix'),
+        ('name_title', 'cmte_custodian_title'),
+        ('zip', 'cmte_custodian_zip'),
+    )
+    custodian = (
+        ('custodian', 'cmte_custodian_city,cmte_custodian_f_nm\
+            ,cmte_custodian_l_nm,cmte_custodian_m_nm,cmte_custodian_nm,\
+            cmte_custodian_ph_num,cmte_custodian_prefix,cmte_custodian_st,\
+            cmte_custodian_st1,cmte_custodian_st2,cmte_custodian_suffix,\
+            cmte_custodian_title,cmte_custodian_zip'),
+    )
+    properties_field_mapping = (
+        ('email', 'cmte_email'),
+        ('fax', 'cmte_fax'),
+        ('name', 'cmte_nm'),
+        ('website', 'cmte_web_url'),
+        ('expire_date', 'expire_date'),
+        ('filing_frequency', 'filing_freq'),
+        ('form_type', 'form_tp'),
+        ('leadership_pac', 'leadership_pac'),
+        ('load_date', 'load_date'),
+        ('lobbyist_registrant_pac', 'lobbyist_registrant_pac_flg'),
+        ('organization_type', 'org_tp'),
+        ('organization_type_full', 'org_tp_desc'),
+        ('original_registration_date', 'orig_registration_dt'),
+        ('party', 'cand_pty_affiliation'),
+        ('party_type', 'party_cmte_type'),
+        ('party_type_full', 'party_cmte_type_desc'),
+        ('qualifying_date', 'qual_dt'),
+    )
+
+    properties_mapping = (
+        properties_field_mapping + committee_address_field_mappings
+        + committee_address + treasurer + treasurer_field_mapping
+        + custodian_field_mapping + custodian + (('*', '*'),))
+
+    # connects mappings to field names
+    maps_fields = (
+        (dimcmte_mapping, 'dimcmte_fields'),
+        (properties_mapping, 'properties_fields'),
+        (linkages_mapping, 'linkages_fields'),
+        (designation_mapping, 'designation_fields'),
+    )
